+++ src/components/premieres/Section/PremiereSection.style.ts
@@ -14,3 +14,25 @@ export const Section = styled.section`
     gap: 24px;
   }
 `;
+
+export const Card = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+
+  @media (min-width: 0px) and (max-width: 600px) {
+    width: 100%;
+    min-width: 148px;
+  }
+
+  @media (min-width: 600px) {
+    width: 168px;
+  }
+`;
+
+export const ThumbnailImage = styled.img`
+  width: 100%;
+  aspect-ratio: '3 / 4';
+  object-fit: cover;
+  border-radius: 4px;
+`;
